Implementation_Phases/Phase_F_Generation/requirements.txt
langchain-groq==0.1.2
python-dotenv==1.0.1
